# Accelerate `CELU` and add tests