--- a/src/mlpack/methods/ann/layer/celu.hpp
+++ b/src/mlpack/methods/ann/layer/celu.hpp
@@ -121,10 +121,7 @@ class CELUType : public Layer<MatType>
   void serialize(Archive& ar, const uint32_t /* version */);
 
  private:
-  //! Locally stored first derivative of the activation function.
-  MatType derivative;
-
-  //! CELU Hyperparameter (alpha > 0).
+  // CELU Hyperparameter (alpha > 0).
   double alpha;
 }; // class CELUType