--- a/src/mlpack/methods/ann/layer/celu_impl.hpp
+++ b/src/mlpack/methods/ann/layer/celu_impl.hpp
@@ -63,13 +63,13 @@ template<typename MatType>
 CELUType<MatType>&
 CELUType<MatType>::operator=(CELUType&& other)
 {
-    if (&other != this)
-    {
-      Layer<MatType>::operator=(std::move(other));
-      alpha = std::move(other.alpha);
-    }
+  if (&other != this)
+  {
+    Layer<MatType>::operator=(std::move(other));
+    alpha = std::move(other.alpha);
+  }
 
-    return *this;
+  return *this;
 }
 
 template<typename MatType>
@@ -81,26 +81,19 @@ void CELUType<MatType>::Forward(
     output(i) = (input(i) >= 0) ? input(i) : alpha *
         (std::exp(input(i) / alpha) - 1);
   }
-
-  if (this->training)
-  {
-    derivative.set_size(arma::size(input));
-    for (size_t i = 0; i < input.n_elem; ++i)
-    {
-      derivative(i) = (input(i) >= 0) ? 1 :
-          (output(i) / alpha) + 1;
-    }
-  }
 }
 
 template<typename MatType>
 void CELUType<MatType>::Backward(
-    const MatType& /* input */,
-    const MatType& /* output */,
+    const MatType& input,
+    const MatType& output,
     const MatType& gy,
     MatType& g)
 {
-  g = gy % derivative;
+  for (size_t i = 0; i < input.n_elem; ++i)
+  {
+    g(i) = gy(i) * ((input(i) >= 0) ? 1 : (output(i) / alpha) + 1);
+  }
 }
 
 template<typename MatType>
@@ -112,8 +105,6 @@ void CELUType<MatType>::serialize(
   ar(cereal::base_class<Layer<MatType>>(this));
 
   ar(CEREAL_NVP(alpha));
-  if (Archive::is_loading::value)
-    derivative.clear();
 }
 
 } // namespace mlpack