I noticed that the `CELU` layer held a matrix of derivatives between `Forward()` and `Backward()`.  But this is really not a great implementation as the output of the layer is available when `Backward()` is called.  So, I removed the member and simplified the layer.  On a simple timing test for `Forward()` and `Backward()` with 1M random elements between -0.5 and 0.5:

before:

```
Forward(): 0.0339329s
Backward(): 0.0239357s
```

after:

```
Forward(): 0.0176554s
Backward(): 0.0161865s
```

and memory usage will be smaller too.  Maybe it is possible to do better, I didn't put too much effort into it.